requirements.txt
mfrc522==0.0.7
RPi.GPIO==0.7.1
spidev==3.8
